start refactoring stuff
